Improved chanks that values don't go out-of-range.

+++ Panel/Input.h
@@ -142,7 +142,7 @@ int        inputNodes  = 0;   // Bit map of Input nodes present.
 int        inputNumber = 0;   // Current Input number.
 InputDef   inputDef;          // Definition of the current Input.
 uint32_t   inputTypes  = 0L;  // The types of the Inputs. 2 bits per pin, 16 pins per node = 32 bits.
-int        inputType   = 0;   // Type of the current Input (2 bits, INPUT_TYPE_MASK).
+uint8_t    inputType   = 0;   // Type of the current Input (2 bits, INPUT_TYPE_MASK).
 
 
 /** Load an Input's data from EEPROM.
